0602 from home 1.1.0.4

+++ SuperMinersServerApplication/SuperMinersServerApplication/WebService/Contracts/IServiceToClient.Game.cs
@@ -33,7 +33,7 @@ namespace SuperMinersServerApplication.WebService.Contracts
             ResponseFormat = WebMessageFormat.Json,
             RequestFormat = WebMessageFormat.Json,
             BodyStyle = WebMessageBodyStyle.WrappedRequest)]
-        int GatherStones(string token, string userName, int stones);
+        int GatherStones(string token, string userName, float stones);
 
         [OperationContract]
         [WebInvoke(UriTemplate = "/WebService/SellStones",
